[move]: ItemOption 클래스 이동
