+++ BE/backend/src/main/java/project/main/webstore/domain/order/dto/OrderItemResponseDto.java
@@ -4,7 +4,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import project.main.webstore.domain.order.enums.OrderedItem;
+import project.main.webstore.domain.order.entity.OrderedItem;
 
 @Getter
 @Setter
